refactor(bridge-flows): extract TransfersTable + row cells to _components/
Move ~460 lines of inline sub-components out of bridge-flows/page.tsx into two
new files under bridge-flows/_components/:

- transfers-table.tsx (~242 lines): TransfersTable with sort state + header
  rendering + tbody row mapping
- transfer-row-cells.tsx (~300 lines): TimeCell, DurationCell, RouteCell,
  Dash, WormholescanLink, TxPill, TxLinks, TokenCell, SenderCell

page.tsx shrinks from 909 → 421 lines. RouteDeliveryTile stays in page.tsx
and imports RouteCell from the new module. All 1687 tests (including the 42
bridge-flows characterization tests) stay green. No behavior changes.
