+++ ui-dashboard/src/app/bridge-flows/page.tsx
@@ -19,59 +19,28 @@ import {
   BRIDGE_DELIVERED_RECENT,
 } from "@/lib/bridge-queries";
 import { TOP_BRIDGERS_EXPANDED } from "@/lib/bridge-flows/layout";
-import {
-  ALL_BRIDGE_STATUSES,
-  deriveBridgeStatus,
-  formatDurationShort,
-  transferDeliveryDurationSec,
-} from "@/lib/bridge-status";
-import { BridgeStatusBadge } from "@/components/bridge-status-badge";
+import { ALL_BRIDGE_STATUSES, formatDurationShort } from "@/lib/bridge-status";
 import { BridgeStatusFilter } from "@/components/bridge-status-filter";
-import { BridgeProviderBadge } from "@/components/bridge-provider-badge";
 import {
-  BridgeRedeemPill,
   ToastPortal,
   type AddToast,
   type ToastEntry,
 } from "@/components/bridge-redeem-cta";
-import { ChainIcon } from "@/components/chain-icon";
 import { Tile, Skeleton, ErrorBox, EmptyBox } from "@/components/feedback";
 import { Pagination } from "@/components/pagination";
 import { BreakdownTile } from "@/components/breakdown-tile";
 import { BridgeVolumeChart } from "@/components/bridge-volume-chart";
 import { BridgeTopBridgersChart } from "@/components/bridge-top-bridgers-chart";
 import { BridgeTokenBreakdownChart } from "@/components/bridge-token-breakdown-chart";
-import { Table, Th } from "@/components/table";
-import { SortableTh } from "@/components/sortable-th";
-import { AddressLink } from "@/components/address-link";
 import { useOracleRates } from "@/hooks/use-oracle-rates";
 import { ENVIO_MAX_ROWS } from "@/lib/constants";
-import {
-  formatWei,
-  formatUSD,
-  relativeTime,
-  formatTimestamp,
-  truncateAddress,
-} from "@/lib/format";
-import { networkForChainId, tokenAddressForSymbol } from "@/lib/networks";
-import {
-  explorerAddressUrl,
-  explorerTxUrl,
-  type OracleRateMap,
-} from "@/lib/tokens";
-import { sortTransfers, type BridgeSortKey } from "@/lib/bridge-flows/sort";
-import {
-  transferAmountUsd,
-  usdPricedFromLiveRate,
-} from "@/lib/bridge-flows/pricing";
 import { windowTotals } from "@/lib/bridge-flows/snapshots";
-import { canManuallyRedeemTransfer } from "@/lib/bridge-flows/redeem";
 import { computeRouteAvgDeliverTimes } from "@/lib/bridge-flows/route-stats";
-import { wormholescanUrl } from "@/lib/wormhole/urls";
+import { TransfersTable } from "./_components/transfers-table";
+import { RouteCell } from "./_components/transfer-row-cells";
 import type {
   BridgeBridger,
   BridgeDailySnapshot,
-  BridgeProvider,
   BridgeStatus,
   BridgeTransfer,
 } from "@/lib/types";
@@ -446,463 +415,6 @@ function RouteDeliveryTile({
             last {ROUTE_STATS_LIMIT} delivered
           </p>
         </>
-      )}
-    </div>
-  );
-}
-
-function TimeCell({ ts, whUrl }: { ts: string | null; whUrl: string | null }) {
-  const relative = ts ? relativeTime(ts) : "—";
-  const precise = ts ? formatTimestamp(ts) : undefined;
-  if (whUrl) {
-    return (
-      <a
-        href={whUrl}
-        target="_blank"
-        rel="noopener noreferrer"
-        title={precise}
-        className="text-slate-400 hover:text-indigo-300 transition-colors"
-      >
-        {relative}
-      </a>
-    );
-  }
-  return (
-    <span className="text-slate-400" title={precise}>
-      {relative}
-    </span>
-  );
-}
-
-function TransfersTable({
-  transfers,
-  rates,
-  addToast,
-}: {
-  transfers: BridgeTransfer[];
-  rates: OracleRateMap;
-  addToast: AddToast;
-}) {
-  const [sortKey, setSortKey] = useState<BridgeSortKey>("time");
-  const [sortDir, setSortDir] = useState<"asc" | "desc">("desc");
-
-  const handleSort = (key: BridgeSortKey) => {
-    if (key === sortKey) {
-      setSortDir((d) => (d === "asc" ? "desc" : "asc"));
-    } else {
-      setSortKey(key);
-      setSortDir("desc");
-    }
-  };
-
-  const sorted = useMemo(
-    () => sortTransfers(transfers, sortKey, sortDir, rates),
-    [transfers, sortKey, sortDir, rates],
-  );
-
-  return (
-    <Table>
-      <thead>
-        <tr className="border-b border-slate-800 bg-slate-900/50">
-          <SortableTh
-            sortKey="provider"
-            activeSortKey={sortKey}
-            sortDir={sortDir}
-            onSort={handleSort}
-          >
-            Provider
-          </SortableTh>
-          <SortableTh
-            sortKey="route"
-            activeSortKey={sortKey}
-            sortDir={sortDir}
-            onSort={handleSort}
-          >
-            Route
-          </SortableTh>
-          <SortableTh
-            sortKey="status"
-            activeSortKey={sortKey}
-            sortDir={sortDir}
-            onSort={handleSort}
-          >
-            Status
-          </SortableTh>
-          <SortableTh
-            sortKey="token"
-            activeSortKey={sortKey}
-            sortDir={sortDir}
-            onSort={handleSort}
-          >
-            Token
-          </SortableTh>
-          <SortableTh
-            sortKey="amountUsd"
-            activeSortKey={sortKey}
-            sortDir={sortDir}
-            onSort={handleSort}
-            align="right"
-          >
-            Amount
-          </SortableTh>
-          <SortableTh
-            sortKey="sender"
-            activeSortKey={sortKey}
-            sortDir={sortDir}
-            onSort={handleSort}
-          >
-            Sender
-          </SortableTh>
-          <SortableTh
-            sortKey="receiver"
-            activeSortKey={sortKey}
-            sortDir={sortDir}
-            onSort={handleSort}
-          >
-            Receiver
-          </SortableTh>
-          <Th>Txs</Th>
-          <SortableTh
-            sortKey="time"
-            activeSortKey={sortKey}
-            sortDir={sortDir}
-            onSort={handleSort}
-            align="right"
-          >
-            Time
-          </SortableTh>
-          <SortableTh
-            sortKey="duration"
-            activeSortKey={sortKey}
-            sortDir={sortDir}
-            onSort={handleSort}
-            align="right"
-          >
-            Duration
-          </SortableTh>
-        </tr>
-      </thead>
-      <tbody>
-        {sorted.map((t) => {
-          const status = deriveBridgeStatus(t);
-          const usd = transferAmountUsd(t, rates);
-          const usdFromLive = usd !== null && usdPricedFromLiveRate(t);
-          const sameParties =
-            !!t.sender &&
-            !!t.recipient &&
-            t.sender.toLowerCase() === t.recipient.toLowerCase();
-          const whUrl =
-            t.provider === "WORMHOLE" && t.sentTxHash
-              ? wormholescanUrl(t.sentTxHash)
-              : null;
-          const redeemProps =
-            status === "STUCK" && canManuallyRedeemTransfer(t)
-              ? {
-                  sentTxHash: t.sentTxHash!,
-                  destChainId: t.destChainId!,
-                  tokenSymbol: t.tokenSymbol,
-                }
-              : null;
-          return (
-            <tr
-              key={t.id}
-              className="border-b border-slate-800/50 hover:bg-slate-800/30 transition-colors"
-            >
-              <td className="px-2 sm:px-3 py-1.5 sm:py-2">
-                <WormholescanLink href={whUrl}>
-                  <BridgeProviderBadge provider={t.provider} />
-                </WormholescanLink>
-              </td>
-              <td className="px-2 sm:px-3 py-1.5 sm:py-2">
-                <RouteCell
-                  sourceChainId={t.sourceChainId}
-                  destChainId={t.destChainId}
-                />
-              </td>
-              <td className="px-2 sm:px-3 py-1.5 sm:py-2">
-                <BridgeStatusBadge status={status} />
-              </td>
-              <td className="px-2 sm:px-3 py-1.5 sm:py-2 text-sm">
-                <TokenCell
-                  symbol={t.tokenSymbol}
-                  chainId={t.sourceChainId ?? t.destChainId}
-                />
-              </td>
-              <td className="px-2 sm:px-3 py-1.5 sm:py-2 font-mono text-right">
-                <WormholescanLink href={whUrl}>
-                  <div
-                    className="text-sm text-slate-200"
-                    title={
-                      usdFromLive
-                        ? "USD priced at render time from current oracle rate"
-                        : undefined
-                    }
-                  >
-                    {usd === null
-                      ? "—"
-                      : `${usdFromLive ? "~" : ""}${formatUSD(usd)}`}
-                  </div>
-                  {t.amount && (
-                    <div className="text-xs text-slate-500">
-                      {formatWei(t.amount, t.tokenDecimals ?? 18, 2)}
-                    </div>
-                  )}
-                </WormholescanLink>
-              </td>
-              <td className="px-2 sm:px-3 py-1.5 sm:py-2 text-sm">
-                <SenderCell sender={t.sender} chainId={t.sourceChainId} />
-              </td>
-              <td
-                className={`px-2 sm:px-3 py-1.5 sm:py-2 text-sm ${sameParties ? "opacity-50" : ""}`}
-                title={sameParties ? "Same as sender" : undefined}
-              >
-                <SenderCell sender={t.recipient} chainId={t.destChainId} />
-              </td>
-              <td className="px-2 sm:px-3 py-1.5 sm:py-2">
-                <TxLinks
-                  provider={t.provider}
-                  sentTxHash={t.sentTxHash}
-                  sourceChainId={t.sourceChainId}
-                  deliveredTxHash={t.deliveredTxHash}
-                  destChainId={t.destChainId}
-                  redeemProps={redeemProps}
-                  addToast={addToast}
-                />
-              </td>
-              <td className="px-2 sm:px-3 py-1.5 sm:py-2 text-xs font-mono text-right whitespace-nowrap">
-                <TimeCell ts={t.sentTimestamp ?? t.firstSeenAt} whUrl={whUrl} />
-              </td>
-              <DurationCell transfer={t} />
-            </tr>
-          );
-        })}
-      </tbody>
-    </Table>
-  );
-}
-
-function DurationCell({ transfer }: { transfer: BridgeTransfer }) {
-  const durationSec = transferDeliveryDurationSec(transfer);
-  if (durationSec === null) {
-    // Mirror the STUCK overlay: any non-terminal-delivered status is
-    // "pending" from a duration perspective — including the client-side
-    // STUCK overlay, which should still show "pending" rather than em-dash
-    // (it's unfinished, not unknown). CANCELLED/FAILED are unreachable on
-    // the bridge page today (no indexer handler writes them) but the
-    // em-dash branch is kept for schema-level safety.
-    const derived = deriveBridgeStatus(transfer);
-    const pending =
-      derived !== "DELIVERED" &&
-      derived !== "CANCELLED" &&
-      derived !== "FAILED";
-    return (
-      <td
-        className="px-2 sm:px-3 py-1.5 sm:py-2 text-xs text-slate-500 font-mono text-right whitespace-nowrap"
-        title={
-          pending
-            ? "Not yet delivered"
-            : "Delivery timestamp unavailable for this transfer"
-        }
-      >
-        {pending ? "pending" : "\u2014"}
-      </td>
-    );
-  }
-  return (
-    <td
-      className="px-2 sm:px-3 py-1.5 sm:py-2 text-xs text-slate-400 font-mono text-right whitespace-nowrap"
-      title="Source-send to destination-delivery elapsed time"
-    >
-      {formatDurationShort(durationSec)}
-    </td>
-  );
-}
-
-function RouteCell({
-  sourceChainId,
-  destChainId,
-}: {
-  sourceChainId: number | null;
-  destChainId: number | null;
-}) {
-  const src = networkForChainId(sourceChainId);
-  const dst = networkForChainId(destChainId);
-  return (
-    <span className="inline-flex items-center gap-1.5 text-xs text-slate-300">
-      {src ? <ChainIcon network={src} size={14} /> : <Dash />}
-      <span className="text-slate-500">{"\u2192"}</span>
-      {dst ? <ChainIcon network={dst} size={14} /> : <Dash />}
-    </span>
-  );
-}
-
-function Dash() {
-  return <span className="text-slate-600">{"\u2014"}</span>;
-}
-
-/** Wraps children in a Wormholescan tx anchor when href is set. Keeps the
- * cell's typography — no blue link recolor on the content. */
-function WormholescanLink({
-  href,
-  children,
-}: {
-  href: string | null;
-  children: React.ReactNode;
-}) {
-  if (!href) return <>{children}</>;
-  return (
-    <a
-      href={href}
-      target="_blank"
-      rel="noopener noreferrer"
-      title="Open on Wormholescan"
-      className="hover:text-indigo-300 transition-colors"
-    >
-      {children}
-    </a>
-  );
-}
-
-function TxPill({
-  href,
-  label,
-  title,
-}: {
-  href: string;
-  label: string;
-  title: string;
-}) {
-  return (
-    <a
-      href={href}
-      target="_blank"
-      rel="noopener noreferrer"
-      title={title}
-      className="inline-flex items-center gap-0.5 rounded bg-slate-800 px-1.5 py-0.5 text-[10px] font-mono text-slate-400 hover:bg-slate-700 hover:text-slate-200 transition-colors"
-    >
-      {label}
-      <span aria-hidden="true" className="text-slate-600">
-        {"\u2197"}
-      </span>
-    </a>
-  );
-}
-
-function TxLinks({
-  provider,
-  sentTxHash,
-  sourceChainId,
-  deliveredTxHash,
-  destChainId,
-  redeemProps,
-  addToast,
-}: {
-  provider: BridgeProvider;
-  sentTxHash: string | null;
-  sourceChainId: number | null;
-  deliveredTxHash: string | null;
-  destChainId: number | null;
-  redeemProps: {
-    sentTxHash: string;
-    destChainId: number;
-    tokenSymbol: string;
-  } | null;
-  addToast: AddToast;
-}) {
-  const src = networkForChainId(sourceChainId);
-  const dst = networkForChainId(destChainId);
-  const pills: Array<{ href: string; label: string; title: string }> = [];
-  if (sentTxHash && src) {
-    pills.push({
-      href: explorerTxUrl(src, sentTxHash),
-      label: "src",
-      title: `Source tx on ${src.label}`,
-    });
-  }
-  if (deliveredTxHash && dst) {
-    pills.push({
-      href: explorerTxUrl(dst, deliveredTxHash),
-      label: "dst",
-      title: `Destination tx on ${dst.label}`,
-    });
-  }
-  // Wormholescan only resolves by source tx hash or VAA ID; digest alone
-  // 404s. Skip the pill when we don't have the source tx yet.
-  if (provider === "WORMHOLE" && sentTxHash) {
-    pills.push({
-      href: wormholescanUrl(sentTxHash),
-      label: "wh",
-      title: "End-to-end trace on Wormholescan",
-    });
-  }
-  if (pills.length === 0 && !redeemProps) return <Dash />;
-  return (
-    <span className="inline-flex items-center gap-1">
-      {pills.map((p) => (
-        <TxPill key={p.label} {...p} />
-      ))}
-      {redeemProps ? (
-        <BridgeRedeemPill {...redeemProps} addToast={addToast} />
-      ) : null}
-    </span>
-  );
-}
-
-function TokenCell({
-  symbol,
-  chainId,
-}: {
-  symbol: string;
-  chainId: number | null;
-}) {
-  // Resolve the per-chain token address from @mento-protocol/contracts via
-  // the network's tokenSymbols map — NOT from the indexer-stored
-  // BridgeTransfer.tokenAddress. The NTT hub/spoke model deploys a distinct
-  // token address per chain, and legacy indexer data (pre-b390cc9) can carry
-  // the destination-chain's address tagged with the source chain id, which
-  // would produce broken cross-chain explorer links. Symbol + chainId is
-  // authoritative and stable across indexer state.
-  const net = networkForChainId(chainId);
-  const address = net ? tokenAddressForSymbol(net, symbol) : null;
-  if (!net || !address) {
-    return <span className="font-mono text-slate-200">{symbol}</span>;
-  }
-  return (
-    <a
-      href={explorerAddressUrl(net, address)}
-      target="_blank"
-      rel="noopener noreferrer"
-      title={address}
-      className="font-mono text-slate-200 hover:text-indigo-300 transition-colors"
-    >
-      {symbol}
-      <span className="ml-1 text-slate-600" aria-hidden="true">
-        {"\u2197"}
-      </span>
-    </a>
-  );
-}
-
-function SenderCell({
-  sender,
-  chainId,
-}: {
-  sender: string | null;
-  chainId: number | null;
-}) {
-  if (!sender) return <Dash />;
-  const net = networkForChainId(chainId);
-  // `<div>` (not `<span>`) because AddressLink can render AddressLabelEditor,
-  // which mounts a `<dialog>` — phrasing content can't contain flow content.
-  return (
-    <div className="inline-flex items-center gap-1.5">
-      {net && <ChainIcon network={net} size={14} />}
-      {chainId ? (
-        <AddressLink address={sender} chainId={chainId} />
-      ) : (
-        <span className="font-mono text-xs text-slate-400">
-          {truncateAddress(sender)}
-        </span>
       )}
     </div>
   );
